Improve error handling, code quality, and download path management
