+++ main.py
@@ -1,518 +1,36 @@
-from typing import TypedDict, List, Tuple, Dict, Optional, Any, Set
-import concurrent.futures
-from threading import Lock
-
-from spotapi import Public
-from innertube import InnerTube
-from yt_dlp import YoutubeDL
-
-from time import sleep
-from random import uniform
+from typing import Optional
 import os
 import sys
-    
+import time
+
+# Import from utility modules
+from spotify_utils import get_playlist_info
+from ytmusic_utils import get_song_urls
+from downloader import download_from_urls, DownloadStats
+
 # Global variables
 DOWNLOAD_PATH = "./downloads/" # ends with "/"
 client = None
 FFMPEG_PATH = None  # Will store custom FFmpeg path if provided
 
 
-class PlaylistInfo(TypedDict):
-    """TypedDict for structured playlist information"""
-    title: str
-    artist: str
-    length: int
-
-
-class SongData(TypedDict):
-    """TypedDict for structured song data"""
-    url: str
-    title: str
-    original_title: str
-    artist: str
-
-
-class DownloadStats(TypedDict):
-    """TypedDict for structured download statistics"""
-    start_time: float
-    total_bytes: int
-    downloaded_bytes: int
-    current_speed: float
-    successful: List[SongData]
-    failed: List[SongData]
-    in_progress: Set[str]
-    completed: int
-
-
-def get_playlist_info(playlist_id: str) -> List[PlaylistInfo]:
-    """Extracts data from Spotify and return them in format
-       `[{"title": title, "artist": artist, "length": length}]`."""
-
-    items = next(Public.playlist_info(playlist_id))["items"]
-
-    result = []
+def extract_playlist_id(playlist_url: str) -> str:
+    """Extract playlist ID from a Spotify playlist URL
     
-    for item in items:
-        song = {}
-        item = item["itemV2"]["data"]
-        song["title"] = item["name"]
-        song["artist"] = item["artists"]["items"][0]["profile"]["name"]
-        song["length"] = int(item["trackDuration"]["totalMilliseconds"])
-        result.append(song)
-
-    return result
-
-def convert_to_milliseconds(text: str) -> int:
-    """Converts `"%M:%S"` timestamp from YTMusic to milliseconds."""
-    minutes, seconds = text.split(":")
-    return (int(minutes) * 60 + int(seconds)) * 1000
-    
-def get_song_url(song_info: PlaylistInfo, client: Optional[InnerTube] = None) -> Tuple[Optional[str], Optional[str]]:
-    """Simulates searching from the YTMusic web and returns url to closest match based on song length."""
-
-    if client is None:
-        client = InnerTube("WEB_REMIX", "1.20250203.01.00")
-    data = client.search(f"{song_info['title']} {song_info['artist']}")
-
-    try:
-        # Handle "did you mean" case
-        if "itemSectionRenderer" in data["contents"]["tabbedSearchResultsRenderer"]["tabs"][0]["tabRenderer"]["content"]["sectionListRenderer"]["contents"][0]:
-            del data["contents"]["tabbedSearchResultsRenderer"]["tabs"][0]["tabRenderer"]["content"]["sectionListRenderer"]["contents"][0]
-
-        # Extract top result length and video info
-        try:
-            top_result_length = data["contents"]["tabbedSearchResultsRenderer"]["tabs"][0]["tabRenderer"]["content"]["sectionListRenderer"]["contents"][0]["musicCardShelfRenderer"]["subtitle"]["runs"][-1]["text"]
-            top_result_video_id = data["contents"]["tabbedSearchResultsRenderer"]["tabs"][0]["tabRenderer"]["content"]["sectionListRenderer"]["contents"][0]["musicCardShelfRenderer"]["title"]["runs"][0]["navigationEndpoint"]["watchEndpoint"]["videoId"]
-            top_result_title = data["contents"]["tabbedSearchResultsRenderer"]["tabs"][0]["tabRenderer"]["content"]["sectionListRenderer"]["contents"][0]["musicCardShelfRenderer"]["title"]["runs"][0]["text"]
-        except (KeyError, IndexError):
-            top_result_length = None
-            top_result_video_id = None
-            top_result_title = None
-
-        # Extract first song result length and video info
-        try:
-            first_song_length = data["contents"]["tabbedSearchResultsRenderer"]["tabs"][0]["tabRenderer"]["content"]["sectionListRenderer"]["contents"][1]["musicShelfRenderer"]["contents"][0]["musicResponsiveListItemRenderer"]["flexColumns"][1]["musicResponsiveListItemFlexColumnRenderer"]["text"]["runs"][-1]["text"]
-            first_song_video_id = data["contents"]["tabbedSearchResultsRenderer"]["tabs"][0]["tabRenderer"]["content"]["sectionListRenderer"]["contents"][1]["musicShelfRenderer"]["contents"][0]["musicResponsiveListItemRenderer"]["overlay"]["musicItemThumbnailOverlayRenderer"]["content"]["musicPlayButtonRenderer"]["playNavigationEndpoint"]["watchEndpoint"]["videoId"]
-            first_song_title = data["contents"]["tabbedSearchResultsRenderer"]["tabs"][0]["tabRenderer"]["content"]["sectionListRenderer"]["contents"][1]["musicShelfRenderer"]["contents"][0]["musicResponsiveListItemRenderer"]["flexColumns"][0]["musicResponsiveListItemFlexColumnRenderer"]["text"]["runs"][0]["text"]
-        except (KeyError, IndexError):
-            first_song_length = None
-            first_song_video_id = None
-            first_song_title = None
-
-        # Compare song lengths to find the best match
-        if top_result_length and first_song_length and top_result_video_id and first_song_video_id:
-            top_result_diff = abs(convert_to_milliseconds(top_result_length) - song_info["length"])
-            first_song_diff = abs(convert_to_milliseconds(first_song_length) - song_info["length"])
-            
-            if top_result_diff < first_song_diff:
-                # Top result is a better match
-                video_id = top_result_video_id
-                video_title = top_result_title
-            else:
-                # First song is a better match
-                video_id = first_song_video_id
-                video_title = first_song_title
-        elif top_result_video_id:
-            # Only top result available
-            video_id = top_result_video_id
-            video_title = top_result_title
-        elif first_song_video_id:
-            # Only first song available
-            video_id = first_song_video_id
-            video_title = first_song_title
-        else:
-            raise ValueError("Could not find video ID in search results")
-
-        url = f"https://music.youtube.com/watch?v={video_id}"
-        return url, video_title
-
-    except Exception as e:
-        print(f"Error processing search results: {str(e)}")
-        # Return a tuple to maintain compatibility
-        return None, None
-
-def get_song_urls(playlist_info: List[PlaylistInfo], progress_callback: Optional[callable] = None, concurrent_searches: int = 3) -> List[SongData]:
-    """Searches for songs in playlist concurrently and returns list of results."""
-    urls = []
-    client = InnerTube("WEB_REMIX", "1.20250203.01.00")
-    total_songs = len(playlist_info)
-    
-    # Mutex for thread-safe updates to the progress
-    progress_lock = Lock()
-    completed_count = 0
-    
-    def search_song(index: int, song_info: PlaylistInfo) -> bool:
-        nonlocal completed_count
+    Args:
+        playlist_url: Full Spotify playlist URL or ID
         
-        # Update progress at start of search
-        with progress_lock:
-            if progress_callback:
-                progress_callback(f"Finding: {song_info['title']} by {song_info['artist']} ({index}/{total_songs})", 
-                                 20 + (completed_count / total_songs) * 20)
-        
-        # Small delay between API calls to avoid rate limiting (reduced from 1-3 seconds to 0.2-0.5 seconds)
-        sleep(uniform(0.2, 0.5))
-        
-        # Perform the search
-        result = get_song_url(song_info, client)
-        
-        # Process results and update progress
-        with progress_lock:
-            completed_count += 1
-            
-            if result[0]:  # Only add if we got a valid URL
-                # Ensure we're using YouTube Music URL
-                url = result[0]
-                if not url.startswith("https://music.youtube.com"):
-                    url = url.replace("https://www.youtube.com", "https://music.youtube.com")
-                
-                urls.append({
-                    "url": url, 
-                    "title": result[1], 
-                    "original_title": song_info['title'], 
-                    "artist": song_info['artist']
-                })
-                
-                if progress_callback:
-                    progress_callback(f"Found: {result[1]} ({index}/{total_songs})", 
-                                     20 + (completed_count / total_songs) * 20)
-            else:
-                if progress_callback:
-                    progress_callback(f"Skipped: Could not find {song_info['title']} ({index}/{total_songs})", 
-                                     20 + (completed_count / total_songs) * 20)
-        
-        return result[0] is not None
-    
-    # Execute searches concurrently
-    with concurrent.futures.ThreadPoolExecutor(max_workers=concurrent_searches) as executor:
-        futures = {}
-        
-        # Submit all search tasks
-        for index, song_info in enumerate(playlist_info, 1):
-            future = executor.submit(search_song, index, song_info)
-            futures[future] = index
-        
-        # Wait for all searches to complete
-        for future in concurrent.futures.as_completed(futures):
-            try:
-                future.result()
-            except Exception as e:
-                print(f"Error in search task: {str(e)}")
-                if progress_callback:
-                    with progress_lock:
-                        progress_callback(f"Error: {str(e)}", 20 + (completed_count / total_songs) * 20)
-    
-    return urls
+    Returns:
+        Extracted playlist ID
+    """
+    if "spotify.com/playlist/" in playlist_url:
+        # Extract ID from URL
+        playlist_id = playlist_url.split("playlist/")[1].split("?")[0]
+        return playlist_id
+    return playlist_url  # Already an ID
 
-def download_from_urls(urls: List[SongData], progress_callback: Optional[callable] = None, concurrent_downloads: int = 3) -> DownloadStats:
-    """Downloads list of songs with yt-dlp with concurrent downloads support"""
-    
-    import concurrent.futures
-    import time
-    from collections import defaultdict
-    import os
-    import platform
-    from pathlib import Path
-    
-    # Track overall download stats
-    download_stats: DownloadStats = {
-        "start_time": time.time(),
-        "total_bytes": 0,
-        "downloaded_bytes": 0,
-        "current_speed": 0,
-        "successful": [],
-        "failed": [],
-        "in_progress": set(),
-        "completed": 0
-    }
-    stats_lock = Lock()
-    
-    # Check if ffmpeg is available
-    ffmpeg_found = False
-    ffmpeg_path = None
-    
-    # First check if a custom path was provided
-    global FFMPEG_PATH
-    if FFMPEG_PATH and os.path.isfile(FFMPEG_PATH):
-        ffmpeg_found = True
-        ffmpeg_path = FFMPEG_PATH
-        if progress_callback:
-            progress_callback(
-                f"Using custom FFmpeg path: {FFMPEG_PATH}", 
-                10,
-                download_stats
-            )
-    else:
-        try:
-            import subprocess
-            import shutil
-            
-            # First check if ffmpeg is in PATH
-            ffmpeg_in_path = shutil.which("ffmpeg")
-            if ffmpeg_in_path:
-                ffmpeg_found = True
-                ffmpeg_path = ffmpeg_in_path
-            else:
-                # Secondary check using platform-specific commands
-                if platform.system() == "Windows":
-                    # Try multiple common installation locations on Windows
-                    common_paths = [
-                        os.path.join(os.environ.get("ProgramFiles", "C:\\Program Files"), "ffmpeg", "bin", "ffmpeg.exe"),
-                        os.path.join(os.environ.get("ProgramFiles(x86)", "C:\\Program Files (x86)"), "ffmpeg", "bin", "ffmpeg.exe"),
-                        os.path.join(os.environ.get("USERPROFILE", "C:\\Users\\User"), "ffmpeg", "bin", "ffmpeg.exe"),
-                        # Add the current directory and its subdirectories
-                        os.path.join(os.getcwd(), "ffmpeg", "bin", "ffmpeg.exe"),
-                        os.path.join(os.getcwd(), "ffmpeg.exe"),
-                        # Add the directory where the script is located
-                        os.path.join(os.path.dirname(os.path.abspath(__file__)), "ffmpeg.exe"),
-                        os.path.join(os.path.dirname(os.path.abspath(__file__)), "ffmpeg", "bin", "ffmpeg.exe"),
-                    ]
-                    
-                    for path in common_paths:
-                        if os.path.isfile(path):
-                            ffmpeg_found = True
-                            ffmpeg_path = path
-                            # Set the global path for future use
-                            FFMPEG_PATH = path
-                            break
-                            
-                    # Last resort: try the where command
-                    if not ffmpeg_found:
-                        try:
-                            result = subprocess.run(["where", "ffmpeg"], capture_output=True, text=True)
-                            if "ffmpeg" in result.stdout.lower():
-                                ffmpeg_path = result.stdout.strip().split('\n')[0]
-                                ffmpeg_found = os.path.isfile(ffmpeg_path)
-                                if ffmpeg_found:
-                                    FFMPEG_PATH = ffmpeg_path
-                        except Exception:
-                            pass
-                else:  # Linux/Mac
-                    try:
-                        result = subprocess.run(["which", "ffmpeg"], capture_output=True, text=True)
-                        if result.stdout.strip():
-                            ffmpeg_path = result.stdout.strip()
-                            ffmpeg_found = os.path.isfile(ffmpeg_path)
-                            if ffmpeg_found:
-                                FFMPEG_PATH = ffmpeg_path
-                    except Exception:
-                        pass
-        except Exception as e:
-            print(f"Error checking for FFmpeg: {str(e)}")
-            ffmpeg_found = False
-            
-    if not ffmpeg_found and progress_callback:
-        progress_callback(
-            "Warning: FFmpeg not found - audio conversion will be skipped. Please install FFmpeg for best quality or place ffmpeg.exe in the application directory.", 
-            10,
-            download_stats
-        )
-    elif ffmpeg_found and progress_callback:
-        progress_callback(
-            f"FFmpeg detected at: {ffmpeg_path} - audio conversion enabled for high quality output.", 
-            10,
-            download_stats
-        )
-        
-    # If FFmpeg was found, add it to the PATH environment variable temporarily
-    if ffmpeg_found and ffmpeg_path:
-        ffmpeg_dir = os.path.dirname(ffmpeg_path)
-        if ffmpeg_dir not in os.environ.get("PATH", ""):
-            os.environ["PATH"] = ffmpeg_dir + os.pathsep + os.environ.get("PATH", "")
-    
-    class ProgressHook:
-        def __init__(self, callback: callable, song_info: SongData, current_idx: int, total_songs: int):
-            self.callback = callback
-            self.current_title = ""
-            self.song_info = song_info
-            self.current_idx = current_idx
-            self.total_songs = total_songs
-            self.last_downloaded_bytes = 0
-            self.last_time = time.time()
-            self.error_reported = False
-            
-        def __call__(self, d: Dict[str, Any]):
-            with stats_lock:
-                if d['status'] == 'downloading':
-                    # Get title if not set
-                    if self.current_title != d.get('info_dict', {}).get('title', ''):
-                        self.current_title = d.get('info_dict', {}).get('title', '')
-                        download_stats["in_progress"].add(self.current_title)
-                    
-                    # Calculate download speed
-                    if "downloaded_bytes" in d and "total_bytes" in d:
-                        # Update total bytes if we have that info
-                        if d["total_bytes"] and self.current_title in download_stats["in_progress"]:
-                            download_stats["total_bytes"] = max(download_stats["total_bytes"], d["total_bytes"])
-                        
-                        # Calculate current speed for this file
-                        now = time.time()
-                        time_diff = now - self.last_time
-                        if time_diff >= 0.5:  # Update every half second
-                            bytes_diff = d["downloaded_bytes"] - self.last_downloaded_bytes
-                            current_speed = bytes_diff / time_diff if time_diff > 0 else 0
-                            
-                            # Update overall download stats
-                            download_stats["downloaded_bytes"] += bytes_diff
-                            download_stats["current_speed"] = current_speed
-                            
-                            # Reset for next calculation
-                            self.last_downloaded_bytes = d["downloaded_bytes"]
-                            self.last_time = now
-                    
-                    # Calculate progress percentage
-                    progress = d.get('downloaded_bytes', 0) / d.get('total_bytes', 1) * 100 if d.get('total_bytes') else 0
-                    
-                    # Calculate ETA
-                    elapsed = time.time() - download_stats["start_time"]
-                    completed_fraction = download_stats["completed"] / len(urls) if urls else 0
-                    if completed_fraction > 0:
-                        eta_seconds = (elapsed / completed_fraction) - elapsed
-                        eta_str = f" - ETA: {int(eta_seconds // 60)}m {int(eta_seconds % 60)}s"
-                    else:
-                        eta_str = " - Calculating..."
-                    
-                    # Format speed in Mbps
-                    speed_mbps = download_stats["current_speed"] / (1024 * 1024) * 8  # Convert bytes/s to Mbps
-                    speed_str = f" - {speed_mbps:.2f} Mbps"
-                    
-                    if self.callback:
-                        self.callback(
-                            f"Downloading: {self.current_title} ({self.current_idx}/{self.total_songs}){speed_str}{eta_str}", 
-                            min(40 + progress * 0.6, 100),
-                            download_stats
-                        )
-                
-                elif d['status'] == 'finished' and self.current_title:
-                    # Remove from in_progress and add to successful
-                    if self.current_title in download_stats["in_progress"]:
-                        download_stats["in_progress"].remove(self.current_title)
-                    download_stats["successful"].append(self.song_info)
-                    
-                    if self.callback:
-                        download_stats["completed"] += 1
-                        overall_progress = (download_stats["completed"] / len(urls)) * 100
-                        
-                        # Calculate ETA
-                        elapsed = time.time() - download_stats["start_time"]
-                        completed_fraction = download_stats["completed"] / len(urls) if urls else 0
-                        if completed_fraction > 0:
-                            eta_seconds = (elapsed / completed_fraction) - elapsed
-                            eta_str = f" - ETA: {int(eta_seconds // 60)}m {int(eta_seconds % 60)}s"
-                        else:
-                            eta_str = " - Calculating..."
-                            
-                        self.callback(
-                            f"Processing: {self.current_title} ({self.current_idx}/{self.total_songs}){eta_str}", 
-                            min(40 + overall_progress * 0.6, 100),
-                            download_stats
-                        )
-                        
-                elif d['status'] == 'error':
-                    # Track failed download and report only once
-                    if not self.error_reported:
-                        if self.current_title in download_stats["in_progress"]:
-                            download_stats["in_progress"].remove(self.current_title)
-                        download_stats["failed"].append(self.song_info)
-                        download_stats["completed"] += 1
-                        self.error_reported = True
-                        
-                        # Report the error (with detailed error message if available)
-                        error_msg = d.get('error', 'Unknown error')
-                        if self.callback:
-                            self.callback(
-                                f"Error: Could not download {self.song_info.get('title', 'Unknown')} - {error_msg}", 
-                                min(40 + (download_stats["completed"] / len(urls) if urls else 0) * 60, 100),
-                                download_stats
-                            )
-    
-    def download_song(song_data: SongData, idx: int, total: int) -> bool:
-        try:
-            # Get the current download path from the global variable
-            current_download_path = DOWNLOAD_PATH
-            
-            # options generated from https://github.com/yt-dlp/yt-dlp/blob/master/devscripts/cli_to_api.py
-            options = {
-                "extract_flat": "discard_in_playlist",
-                "final_ext": "m4a" if ffmpeg_found else "webm",  # Use webm if FFmpeg not available
-                "format": "bestaudio/best",
-                "fragment_retries": 10,
-                "ignoreerrors": "only_download",
-                "outtmpl": {"default": f"{current_download_path}%(title)s.%(ext)s"},
-                "postprocessors": [],  # Will be populated conditionally below
-                "quiet": False,  # Make yt-dlp show output
-                "no_warnings": False,  # Show warnings
-                "retries": 10,
-                "writethumbnail": True,
-                "noplaylist": True,
-                "extractor_args": {"youtubetab": {"approximate_date": "0"}}
-            }
-            
-            # Only add postprocessors if ffmpeg is available
-            if ffmpeg_found:
-                options["postprocessors"] = [
-                    {
-                        "key": "FFmpegExtractAudio",
-                        "nopostoverwrites": False,
-                        "preferredcodec": "m4a",
-                        "preferredquality": "5"
-                    },
-                    {
-                        "add_chapters": True,
-                        "add_infojson": 'if_exists',
-                        "add_metadata": True,
-                        "key": "FFmpegMetadata"
-                    },
-                    {
-                        "key": "EmbedThumbnail",
-                        "already_have_thumbnail": False
-                    }
-                ]
-            
-            if progress_callback:
-                options["progress_hooks"] = [ProgressHook(progress_callback, song_data, idx, total)]
-            
-            # Ensure the URL is a YouTube Music URL
-            url = song_data["url"]
-            if not url.startswith("https://music.youtube.com"):
-                url = url.replace("https://www.youtube.com", "https://music.youtube.com")
-            
-            with YoutubeDL(options) as ydl:
-                ydl.download([url])
-            return True
-        except Exception as e:
-            print(f"Error downloading {song_data.get('title', 'Unknown')}: {str(e)}")
-            with stats_lock:
-                download_stats["failed"].append(song_data)
-                download_stats["completed"] += 1
-            if progress_callback:
-                progress_callback(
-                    f"Error: Could not download {song_data.get('title', 'Unknown')} ({idx}/{total})", 
-                    min(40 + (download_stats["completed"] / len(urls) if urls else 0) * 60, 100),
-                    download_stats
-                )
-            return False
-    
-    total_songs = len(urls)
-    
-    # Use ThreadPoolExecutor for concurrent downloads
-    with concurrent.futures.ThreadPoolExecutor(max_workers=concurrent_downloads) as executor:
-        futures = []
-        for idx, song_data in enumerate(urls, 1):
-            future = executor.submit(download_song, song_data, idx, total_songs)
-            futures.append(future)
-        
-        # Wait for all downloads to complete
-        for future in concurrent.futures.as_completed(futures):
-            try:
-                future.result()
-            except Exception as e:
-                print(f"Unhandled exception: {str(e)}")
-    
-    # Return download stats for reporting
-    return download_stats
 
-def main(playlist_id: str, progress_callback: Optional[callable] = None, concurrent_searches: int = 3, concurrent_downloads: int = 3, ffmpeg_path: Optional[str] = None) -> DownloadStats:
+def main(playlist_id: str, progress_callback: Optional[callable] = None, concurrent_searches: int = 3, concurrent_downloads: int = 3, ffmpeg_path: Optional[str] = None, download_path: Optional[str] = None) -> DownloadStats:
     """Main function to handle the playlist download workflow
     
     Args:
@@ -521,6 +39,7 @@ def main(playlist_id: str, progress_callback: Optional[callable] = None, concurr
         concurrent_searches: Number of concurrent song searches (default: 3)
         concurrent_downloads: Number of concurrent downloads (default: 3)
         ffmpeg_path: Optional path to FFmpeg executable
+        download_path: Optional path to download directory
     
     Returns:
         Download statistics dictionary
@@ -530,16 +49,57 @@ def main(playlist_id: str, progress_callback: Optional[callable] = None, concurr
     if ffmpeg_path and os.path.isfile(ffmpeg_path):
         FFMPEG_PATH = ffmpeg_path
         
+    # Set custom download path if provided
+    global DOWNLOAD_PATH
+    if download_path:
+        DOWNLOAD_PATH = download_path
+        
+    # Extract playlist ID if a full URL was provided
+    playlist_id = extract_playlist_id(playlist_id)
+    
     # Extract playlist info
     playlist_info = get_playlist_info(playlist_id)
+    
+    if not playlist_info:
+        if progress_callback:
+            progress_callback("No songs found in playlist or playlist not found", 100, {
+                "start_time": time.time(),
+                "total_bytes": 0,
+                "downloaded_bytes": 0,
+                "current_speed": 0,
+                "successful": [],
+                "failed": [],
+                "in_progress": set(),
+                "completed": 0
+            })
+        return {
+            "start_time": time.time(),
+            "total_bytes": 0,
+            "downloaded_bytes": 0,
+            "current_speed": 0,
+            "successful": [],
+            "failed": [],
+            "in_progress": set(),
+            "completed": 0
+        }
+    
+    # Update progress
+    if progress_callback:
+        progress_callback(f"Found {len(playlist_info)} songs in playlist", 10)
     
     # Get YouTube Music URLs for each song
     download_urls = get_song_urls(playlist_info, progress_callback, 
                                 concurrent_searches=concurrent_searches)
     
+    # Update progress
+    if progress_callback:
+        progress_callback(f"Found {len(download_urls)} out of {len(playlist_info)} songs", 40)
+    
     # Download songs from the URLs
     download_stats = download_from_urls(download_urls, progress_callback,
-                                      concurrent_downloads=concurrent_downloads)
+                                      concurrent_downloads=concurrent_downloads,
+                                      ffmpeg_path=FFMPEG_PATH,
+                                      download_path=DOWNLOAD_PATH)
     
     # Print final stats
     if progress_callback:
